+++ backend/connector-integration/src/connectors/adyen/test.rs
@@ -152,6 +152,7 @@ mod tests {
                     merchant_config_currency: None,
                     all_keys_required: None,
                     access_token: None,
+                    customer_acceptance: None,
                 },
                 response: Err(ErrorResponse::default()),
             };
@@ -281,6 +282,7 @@ mod tests {
                     merchant_config_currency: None,
                     all_keys_required: None,
                     access_token: None,
+                    customer_acceptance: None,
                 },
                 response: Err(ErrorResponse::default()),
             };